backend/requirements/base.txt
asgiref==3.8.1
astroid==3.3.8
asttokens==3.0.0
black==24.10.0
cachetools==5.5.0
certifi==2024.12.14
chardet==5.2.0
click==8.1.8
colorama==0.4.6
coverage==7.6.10
decorator==5.1.1
dill==0.3.9
distlib==0.3.9
Django==5.1.4
django-debug-toolbar==4.4.6
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
executing==2.1.0
factory_boy==3.3.1
Faker==33.3.1
filelock==3.16.1
h11==0.14.0
inflection==0.5.1
ipython==8.31.0
isort==5.13.2
jedi==0.19.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.10.1
pylint==3.3.3
pyproject-api==1.8.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tomlkit==0.13.2
tox==4.23.2
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.28.1
wcwidth==0.2.13
